Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -7,8 +7,8 @@ char* welcome() {
 int main() {
     int a,b,c;
     scanf("%d %d %d",&a,&b,&c);
-    int avg;
+    float avg;
     avg=(a+b+c)/3;
-    printf("Average: %d ",avg);
+    printf("Average: %f ",avg);
     return 0;
 }
